requirements.txt
django==1.6
djangorestframework==2.3.13
python-social-auth==0.1.23
requests==2.2.1
Pint==0.4.2

coveralls==0.4.1
mock==1.0
nose==1.3
django-nose
pinocchio==0.4.1
markdown
django-filter

